Refactor Object bloomFilterFalseRatio to double

+++ src/java/org/apache/cassandra/tools/nodetool/stats/StatsTableComparator.java
@@ -139,8 +139,8 @@ public class StatsTableComparator implements Comparator<StatsTable>
         }
         else if (sortKey.equals("bloom_filter_false_ratio"))
         {
-            result = compareDoubles((Double) stx.bloomFilterFalseRatio,
-                                    (Double) sty.bloomFilterFalseRatio);
+            result = compareDoubles(stx.bloomFilterFalseRatio,
+                                    sty.bloomFilterFalseRatio);
         }
         else if (sortKey.equals("bloom_filter_off_heap_memory_used"))
         {
